feat(backend): user can register Incomes and Expenses
User can register a currency which can be a Income or an Expenses depending on the Currency
atributte called "expense". If "expense" is true than this currency is an expense. I created a
single Model and Table but 3 different Controllers. CUrrencyController to register new currencies.
ExpenseController to list expenses and CurrencyController to list currencies.